Move the remaining files in the top level build directory
