+++ distributed/__init__.py
@@ -4,6 +4,8 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
+import distributed.model_dist
+
 from distributed.checkpoint import load_checkpoints_to_model
 from distributed.logging_utils import logger
 from distributed.parallel_config import ParallelDims
