+++ generate.py
@@ -26,7 +26,7 @@ from torchchat.cli.builder import (
     TokenizerArgs,
 )
 from torchchat.model import Transformer
-from build.utils import device_sync, set_precision
+from torchchat.utils.build_utils import device_sync, set_precision
 from torchchat.cli.cli import add_arguments_for_verb, arg_init, check_args
 from utils.device_info import get_device_info
