+++ torchchat/usages/openai_api.py
@@ -15,7 +15,7 @@ from typing import Any, Dict, List, Optional, Union
 
 import torch
 
-from build.utils import device_sync
+from torchchat.utils.build_utils import device_sync
 from torchchat.cli.download import is_model_downloaded, load_model_configs
 from generate import Generator, GeneratorArgs
